fixed distance getter being a lil interger bitch and updated t_point to hold double positions (fixed images being avergaed out on the x axis)

+++ includes/cube.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   cube.h                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: fghysbre <fghysbre@stduent.s19.be>         +#+  +:+       +#+        */
+/*   By: fghysbre <fghysbre@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/18 22:56:16 by fghysbre          #+#    #+#             */
-/*   Updated: 2024/10/31 13:36:48 by fghysbre         ###   ########.fr       */
+/*   Updated: 2024/11/01 15:06:52 by fghysbre         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -57,13 +57,13 @@ typedef	struct s_plyr
 
 typedef struct s_point
 {
-	int	x;
-	int	y;
+	double	x;
+	double	y;
 }	t_point;
 
 typedef struct s_ray
 {
-	int		dist;
+	double		dist;
 	t_point	point;
 	int		side;
 	float	rot;
